[mlir][gpu] Improve `gpu-lower-to-nvvm-pipeline` Documentation
This PR improves the documentation for the `gpu-lower-to-nvvm-pipeline` (as it was remaning item for #75775)

- Changes pipeline `gpu-lower-to-nvvm` -> `gpu-lower-to-nvvm-pipeline`
- Adds a section in GPU Dialect in website. It clarifies the pipeline's functionality in lowering primary dialects to NVVM targets.
